Expected merge conflicts in 

- gcc_cross.mak 
  changes in CROSS_COMPILE cause the conflicts: needs manual changes.  
- README.md
  same as for gcc_cross.mak

+++ devices/MIMX94/sm/dev_sm_voltage.c
@@ -215,8 +215,7 @@ int32_t DEV_SM_VoltageLevelSet(uint32_t domainId, int32_t voltageLevel)
             && (voltageLevel <= range->highestVolt))
         {
             /* Change level */
-            status = BRD_SM_SupplyLevelSet(domainId,
-                (uint32_t) voltageLevel);
+            status = BRD_SM_SupplyLevelSet(domainId, voltageLevel);
         }
         else
         {
@@ -238,7 +237,7 @@ int32_t DEV_SM_VoltageLevelSet(uint32_t domainId, int32_t voltageLevel)
 int32_t DEV_SM_VoltageLevelGet(uint32_t domainId, int32_t *voltageLevel)
 {
     int32_t status = SM_ERR_SUCCESS;
-    uint32_t microVolt = 0U;
+    int32_t microVolt = 0;
 
     /* Check domain */
     if (domainId < DEV_SM_NUM_VOLT)
@@ -254,7 +253,7 @@ int32_t DEV_SM_VoltageLevelGet(uint32_t domainId, int32_t *voltageLevel)
     /* Return voltage */
     if (status == SM_ERR_SUCCESS)
     {
-        *voltageLevel = (int32_t) microVolt;
+        *voltageLevel = microVolt;
     }
 
     /* Return status */
